Added tests for import_users_friends.rb.

+++ app/interactions/users/import_users_friends.rb
@@ -51,10 +51,7 @@ class Users::ImportUsersFriends < Less::Interaction
 
   def remove_local_friends
     Friend.all.each do |friend|
-      unless remote_friends_usernames.include?(friend.username)
-        puts "#{friend.username} was removed because you stopped follow them"
-        friend.destroy
-      end
+      friend.destroy unless remote_friends_usernames.include?(friend.username)
     end
   end
